chore: rename wui-tab to wui-tab-item
