+++ apps/gallery-new/stories/composites/wui-tab-item.stories.ts
@@ -1,13 +1,13 @@
 import type { Meta } from '@storybook/web-components'
-import '@reown/appkit-ui/src/composites/wui-tab'
-import type { WuiTab } from '@reown/appkit-ui/src/composites/wui-tab'
+import '@reown/appkit-ui/src/composites/wui-tab-item'
+import type { WuiTab } from '@reown/appkit-ui-new/src/composites/wui-tab-item'
 import { html } from 'lit'
 import { iconOptions } from '../../utils/PresetUtils'
 
 type Component = Meta<WuiTab>
 
 export default {
-  title: 'Composites/wui-tab',
+  title: 'Composites/wui-tab-item',
   args: {
     icon: 'mobile',
     size: 'md',
@@ -34,10 +34,10 @@ export default {
 
 export const Default: Component = {
   render: args =>
-    html`<wui-tab
+    html`<wui-tab-item
       icon=${args.icon}
       size=${args.size}
       label=${args.label}
       ?active=${args.active}
-    ></wui-tab>`
+    ></wui-tab-item>`
 }
