+++ packages/ui-new/index.ts
@@ -43,7 +43,7 @@ export * from './src/composites/wui-otp/index.js'
 export * from './src/composites/wui-qr-code/index.js'
 export * from './src/composites/wui-search-bar/index.js'
 export * from './src/composites/wui-snackbar/index.js'
-export * from './src/composites/wui-tab/index.js'
+export * from './src/composites/wui-tab-item/index.js'
 export * from './src/composites/wui-alertbar/index.js'
 export * from './src/composites/wui-tabs/index.js'
 export * from './src/composites/wui-token-button/index.js'
